chore: Rewrite CLI logic with TS to emit compile
